Fix ckanserviceprovider dependency
Hopefully the tests will pass on Travis now.

Instead of trying to depend on ckanserviceprovider's stable branch from
github in requirements.txt and its master branch in requirements-dev.txt
(which doesn't actually work) just depend on the new 0.0.1 release on
pypi.

To upgrade datapusher to a new version of ckanserviceprovider you
should:

- Publish the new version of ckanserviceprovider to pypi with a new
  version number

- Update the version number for the ckanserviceprovider dependency in
  datapusher's requirements.txt

- Run the datapusher's tests, fix anything that broke

- Push the requirements.txt change and any other changes to datapusher's
  master branch, the changes will get into the next datapusher release