Use ckan.datapusher.callback_url_base when downloading data
The `ckan.datapusher.callback_url_base` config option can be used when
the host where DataPusher is running can not access the public CKAN
site URL. It defines an alternative internal host that will be used in
the API calls to CKAN that the DataPusher performs.
But this was not applied when getting the actual data from an uploaded
resource to CKAN, which has a CKAN url. This changes replaces the host
in the resource URL if the resource is an upload and a value for
`ckan.datapusher.callback_url_base` was provided